Change scaling method in `vdbVolumeToSimpleVolumeImpl`.
Previously, the interval `[min; max]` was rescaled to the interval of the target integer type, where `min` and `max` are the actual minimum and maximum values of the data. This means, that the real absolute values of voxels is lost (as it is spanned into full target type interval anyway).

I propose to always scale `[0; 1]` to the target interval, which should better preserve the values of voxels.