Detect developer builds automatically (#5296)
For developer builds all files are placed in the same directory
(<build dir>/bin) which conflicts with the package file structure for
macOS and Linux. Previously we set a special env variable for this case.
This commit allows to detect it automatically based on relative
locations of the executable and MRMesh library.